% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{billy_goat}
\alias{billy_goat}
\title{A helper function to get the bridgesampling objects}
\usage{
billy_goat(rcfit, ...)
}
\arguments{
\item{rcfit}{the output of the}

\item{...}{options passed directly to the bridge sampler function}
}
\description{
A helper function to get the bridgesampling objects
}
