% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit_to_dta}
\alias{fit_to_dta}
\title{Fit a model, flatten it, and write to dta}
\usage{
fit_to_dta(
  infile,
  outfile = "post.dta",
  mod_num = 1,
  stan_file = NA,
  stan_opts = list()
)
}
\description{
Fit a model, flatten it, and write to dta
}
