% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_stan_code}
\alias{get_stan_code}
\title{Get the raw stan code}
\usage{
get_stan_code(mod_num = 1)
}
\arguments{
\item{mod}{the name of the PWF function}

\item{bhm}{boolean, TRUE indicating to use the BHM model}
}
\description{
Get the raw stan code
}
