% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_stan}
\alias{run_stan}
\title{Esimate the Stan model}
\usage{
run_stan(dat, covars, fname, stan_opts = list(), diag = FALSE)
}
\arguments{
\item{dat}{the name of the stan file}

\item{covars}{the stata-style string for covars}

\item{fname}{the file name of the stan model file}

\item{stan_opts}{the list of options to pass directly to stan}

\item{diag}{TRUE to run diagnostics}
}
\description{
Esimate the Stan model
}
